couple of changes
block.ts:
* move insertTimelineIntoCurrentNote to commands.ts
* move handleColor to colors.ts
* move buildVerticalTimeline to new file timelines/vertical.ts
* move buildHorizontalTimeline to new file timelines/horizontal.ts
* move showEmptyTimelineMessage to new file timelines/index.ts

commands.ts:
* adjust constructor to accept params that enables insertTimelineIntoCurrentNote run correctly
* rearrange so like functions are near each other

main.ts:
* pass new param to TimelineCommandProcessor constructor

timelines/index.ts,
timelines/horizontal.ts,
timelines/vertical.ts:
* created

colors.ts:
* removed handleDynamicColor as an export since handleColor is now in the same file

styles/vertical-timeline.scss:
* made sure anything that was calculating something used calc(...)
