+++ src/block.ts
@@ -1,11 +1,9 @@
-import { AllNotesData, CardContainer, EventItem, InternalTimelineArgs, TimelinesSettings } from './types'
 import type { TFile, MetadataCache, Vault } from 'obsidian'
-import { MarkdownView } from 'obsidian'
-import { RENDER_TIMELINE } from './constants'
+
+import { buildHorizontalTimeline, buildVerticalTimeline, showEmptyTimelineMessage } from './timelines'
+import { AllNotesData, CardContainer, InternalTimelineArgs, TimelinesSettings } from './types'
 import {
-  availableColors,
   buildTimelineDate,
-  createInternalLinkOnNoteCard,
   createTagList,
   convertEntryToMilliseconds,
   filterMDFiles,
@@ -13,17 +11,12 @@ import {
   getEventsInFile,
   getImgUrl,
   getNumEventsInFile,
-  handleDynamicColor,
   isHTMLElementType,
   logger,
   normalizeDate,
   setDefaultArgs,
   sortTimelineDates,
 } from './utils'
-
-// Horizontal (Vis-Timeline) specific imports
-import { Timeline } from 'vis-timeline/esnext'
-import { DataSet } from 'vis-data'
 
 export class TimelineBlockProcessor {
   appVault: Vault
@@ -43,42 +36,6 @@ export class TimelineBlockProcessor {
   setup() {
     this.args = setDefaultArgs()
     this.files = this.appVault.getMarkdownFiles()
-  }
-
-  /**
-   * Insert the statically generated timeline into the current note
-   *
-   * @param sourceView
-   */
-  async insertTimelineIntoCurrentNote(
-    sourceView: MarkdownView,
-  ) {
-    const editor = sourceView.editor
-    if ( !editor ) return
-
-    const source = editor.getValue()
-    const match = RENDER_TIMELINE.exec( source )
-    if ( !match || match.length === 1 ) return
-
-    const tagList = `tags=${match[1]}`
-    logger( 'taglist', tagList )
-
-    const div = document.createElement( 'div' )
-    await this.run( tagList, div )
-
-    const renderedString = `<div class="timeline-rendered">${new Date().toString()}</div>`
-    const rendered = ( new DOMParser()).parseFromString( renderedString, 'text/html' ).body.firstChild
-    div.appendChild( rendered )
-
-    const firstCommentEndIndex = source.indexOf( '-->' )
-    const lastCommentStartIndex = source.lastIndexOf( '<!--' )
-
-    editor.replaceRange(
-      ( new XMLSerializer()).serializeToString( div ),
-      { ch: firstCommentEndIndex + 2, line: 1 },
-      { ch: lastCommentStartIndex - 1, line: 1 },
-      source
-    )
   }
 
   /**
@@ -228,332 +185,12 @@ export class TimelineBlockProcessor {
     }
   }
 
-  handleColor( color: string, noteCard: HTMLDivElement, id: string ): boolean {
-    if ( !availableColors.includes( color )) {
-      handleDynamicColor( color, id )
-      return false
-    }
-
-    noteCard.addClass( color )
-    return true
-  }
-
-  /**
-   * Build a vertical timeline
-   *
-   * @param timeline - the timeline html element
-   * @param timelineNotes - notes which have our timeline tags
-   * @param timelineDates - dates we parsed from event data
-   * @param el - the element to append the timeline to
-   */
-  async buildVerticalTimeline(
-    timeline:HTMLElement,
-    timelineNotes: AllNotesData,
-    timelineDates: string[],
-    el: HTMLElement
-  ) {
-    let eventCount = 0
-    // Build the timeline html element
-    for ( const date of timelineDates ) {
-      const align   = eventCount % 2 === 0 ? 'left' : 'right'
-      const datedTo = [ timelineNotes[date][0].startDate.replace( /-0*$/g, '' )
-                    + ( timelineNotes[date][0].era ? ` ${timelineNotes[date][0].era}` : '' )
-      ]
-      const noteDivs = [timeline.createDiv({
-        cls: ['timeline-container', `timeline-${align}`]
-      }, div => {
-        div.style.setProperty( '--timeline-indent', '0' )
-        div.setAttribute( 'timeline-date', timelineNotes[date][0].startDate )
-        div.setAttribute( 'collapsed', String( false ))
-      })]
-      const eventsDiv = noteDivs[0].createDiv({
-        cls: 'timeline-event-list',
-        attr: { 'style': 'display: block' }
-      })
-      const noteHdrs = [noteDivs[0].createEl( 'h2', {
-        attr: { 'style' : `text-align: ${align};` },
-        text: datedTo[0]
-      })]
-      for ( const note of timelineNotes[date] ) {
-        if ( /^(background|range)$/.test( note.type ) && note.endDate > note.startDate ) {
-          if ( !noteDivs[0].classList.contains( 'timeline-head' )) noteDivs[0].classList.add( 'timeline-head' )
-          const dated = note.endDate + ( note.era ? ` ${note.era}` : '' )
-          if( !datedTo[1] || note.endDate > +noteDivs.last().getAttribute( 'timeline-date' )) {
-            datedTo[1] = datedTo[0] + ' to ' + dated
-          }
-          const noteDiv = timeline.createDiv({
-            cls: ['timeline-container', `timeline-${align}`, 'timeline-tail']
-          }, div => {
-            div.style.setProperty( '--timeline-indent', '0' )
-            div.setAttribute( 'timeline-date', note.endDate )
-          })
-          const noteHdr = noteDiv.createEl( 'h2', {
-            attr: { 'style' : `text-align: ${align};` },
-            text: dated
-          })
-          noteHdrs.push( noteHdr )
-          noteDivs.push( noteDiv )
-          for( const n of noteDivs ) {
-            if( +n.getAttribute( 'timeline-date' ) > note.endDate ) {
-              n.before( noteDiv )
-            }
-          }
-
-          /* The CSS 'timeline-timespan-length' determines both the length of the event timeline widget on the timeline,
-             and the length of the vertical segment that spans between the displayed dates. If this value is not recalculated
-             then these elements will not be responsive to layout changes. */
-          ( noteDivs.last() as HTMLDivElement & { calcLength?: () => void }).calcLength = () =>  {
-            const axisMin = noteDivs[0].getBoundingClientRect().top
-            const axisMax = noteDiv    .getBoundingClientRect().top
-            const spanMin = noteHdrs[0].getBoundingClientRect().top
-            const spanMax = noteHdr    .getBoundingClientRect().top
-            noteDivs[0].style.setProperty( '--timeline-span-length', `${axisMax - axisMin}px` )
-            noteDiv    .style.setProperty( '--timeline-span-length', `${spanMax - spanMin}px` )
-          }
-        }
-      }
-
-      noteDivs[0].addEventListener( 'click', ( event ) => {
-        event.preventDefault()
-        const collapsed = !JSON.parse( noteDivs[0].getAttribute( 'collapsed' ))
-        noteDivs[0].setAttribute( 'collapsed', String( collapsed ))
-        for( const p of noteDivs[0].getElementsByTagName( 'p' )) {
-          p.setCssProps({ 'display': collapsed ? 'none' : 'block' })
-        }
-        /* If this event has a duration (and thus has an end note), we hide all elements between the start and end
-           note along with the end note itself */
-        if( noteDivs.length > 0 ) {
-          noteHdrs[0].setText( collapsed ? datedTo[1] : datedTo[0] )
-          const notes = [...timeline.children]
-          const inner = notes.slice( notes.indexOf( noteDivs.first()) + 1, notes.indexOf( noteDivs.last()) + 1 )
-          inner.forEach(( note: HTMLDivElement & { calcLength?: () => void }) => {
-            note.style.display = collapsed ? 'none' : 'block'
-            note.calcLength?.()
-          })
-        }
-
-        /* The CSS '--timeline-indent' variable allows for scaling down the event when contained within another event,
-           but in this case it also tells us how many time spanning events have had their length altered as a consequence
-           of this note's mutation. */
-        let nested = +noteDivs[0].style.getPropertyValue( '--timeline-indent' ) + noteDivs.length - 1
-        for( let f = 'nextElementSibling', sibling = noteDivs[0][f]; nested > 0; sibling = sibling[f] ) {
-          if( sibling.classList.contains( 'timeline-tail' )) {
-            sibling.calcLength?.()
-            nested--
-          }
-        }
-      })
-
-
-      /*noteContainer[0].addEventListener( 'click', ( event ) => {
-        event.preventDefault()
-        const currentStyle = eventContainer.style
-        if ( currentStyle.getPropertyValue( 'display' ) === 'none' ) {
-          currentStyle.setProperty( 'display', 'block' )
-          return
-        }
-
-        // note from https://github.com/Darakah/obsidian-timelines/pull/58
-        // TODO: Stop Propagation: don't close timeline-card when clicked.
-        // `vis-timeline-graph2d.js` contains a method called `_updateContents` that makes the display
-        // attribute disappear on click via line 7426: `element.innerHTML = '';`
-        currentStyle.setProperty( 'display', 'none' )
-      })*/
-
-      if ( !timelineNotes[date] ) continue
-
-      for ( const eventAtDate of timelineNotes[date] ) {
-        const noteCard = eventsDiv.createDiv({ cls: 'timeline-card' })
-        // add an image only if available
-        if ( eventAtDate.img ) {
-          noteCard.createDiv({
-            cls: 'thumb',
-            attr: { style: `background-image: url(${eventAtDate.img});` }
-          })
-        }
-
-        if ( eventAtDate.color ) {
-          // todo : add more support for other custom classes
-          this.handleColor( eventAtDate.color, noteCard, eventAtDate.id )
-        }
-
-        createInternalLinkOnNoteCard( eventAtDate, noteCard )
-        eventAtDate.body && noteCard.createEl( 'p', { text: eventAtDate.body.trim() })
-      }
-
-      /* If the previous note is of class 'timeline-tail' then there's a chance that the current node belongs nested within
-         the previous one. Here we iterate backwards for as long as elements of the class 'timeline-tail' are encountered.
-
-         Then we sort an array containing the start and ending date of the current event along with the ending date of the
-         previous event. If the index of the ending date of the previous event in the new array is greater than 0 then the
-         note(s) preceding it (either the note belonging to the start date or the notes belonging to both the start and end
-         dates of the current event) are placed before their predecessor. */
-      for( let s = [...timeline.children],i = s.indexOf( noteDivs[0] ); s[i-1]?.classList.contains( 'timeline-tail' ); i-- ) {
-        const t = s[i-1].getAttribute( 'timeline-date' )
-        const times = [t, ...noteDivs.map( n => {
-          return n.getAttribute( 'timeline-date' )
-        })]
-        for( let j = times.sort().lastIndexOf( t ); j > 0; s[i-1].before( ...noteDivs.slice( 0, j )), j = 0 ) {
-          const indent = +noteDivs[0].style.getPropertyValue( '--timeline-indent' ) + 1
-          noteDivs.forEach( n => {
-            n.style.setProperty( '--timeline-indent', `${ indent }` )
-          })
-        }
-      }
-      eventCount++
-    }
-    // Replace the selected tags with the timeline html
-    el.appendChild( timeline )
-
-    // Initial length calculation must be done after appending notes to document
-    el.querySelectorAll( '.timeline-tail' ).forEach(( note: HTMLDivElement & { calcLength?: () => void }) => {
-      note.calcLength?.()
-    })
-    return
-  }
-
-  /**
-   * Build a horizontal timeline
-   *
-   * @param timelineDiv - the timeline html element
-   * @param timelineNotes - notes which have our timeline tags
-   * @param timelineDates - dates we parsed from event data
-   * @param el - the element to append the timeline to
-   */
-  async buildHorizontalTimeline(
-    timelineDiv: HTMLElement,
-    timelineNotes: AllNotesData,
-    timelineDates: string[],
-    el: HTMLElement
-  ) {
-    // Create a DataSet
-    const items = new DataSet( [] )
-
-    if ( !timelineDates ) {
-      logger( 'No dates found for the timeline' )
-      return
-    }
-
-    timelineDates.forEach(( date ) => {
-      // add all events at this date
-      Object.values( timelineNotes[date] ).forEach(( event: CardContainer ) => {
-        const noteCard = document.createElement( 'div' )
-        noteCard.className = 'timeline-card'
-
-        // add an image only if available
-        if ( event.img ) {
-          noteCard.createDiv({
-            cls: 'thumb',
-            attr: { style: `background-image: url(${event.img});` }
-          })
-        }
-
-        let colorIsClass = false
-        if ( event.color ) {
-          colorIsClass = this.handleColor( event.color, noteCard, event.id )
-        }
-
-        createInternalLinkOnNoteCard( event, noteCard )
-        noteCard.createEl( 'p', { text: event.body })
-
-        const start = buildTimelineDate( event.startDate, parseInt( this.settings.maxDigits ))
-        const end = buildTimelineDate( event.endDate, parseInt( this.settings.maxDigits ))
-
-        if (
-          start.toString() === 'Invalid Date' ||
-          ( [ 'range', 'background' ].includes( event.type ) && end.toString() === 'Invalid Date' )
-        ) {
-          console.warn( 'Invalid start or end date - check for Month/Day values that are 0', { start, end, event })
-
-          return
-        }
-
-        const eventItem: EventItem = {
-          id: items.length + 1,
-          content: event.title ?? '',
-          start: start,
-          className: colorIsClass ? event.color ?? 'gray' : `nid-${event.id}`,
-          type: event.type,
-          end: end ?? null,
-          path: event.path,
-          _event: event,
-        }
-
-        // Add Event data
-        items.add( eventItem )
-      })
-    })
-
-    // Configuration for the Timeline
-    const options = {
-      start: this.args.startDate,
-      end: this.args.endDate,
-      min: this.args.minDate,
-      max: this.args.maxDate,
-      minHeight: this.args.divHeight,
-      showCurrentTime: false,
-      showTooltips: false,
-      zoomMin: this.args.zoomInLimit,
-      zoomMax: this.args.zoomOutLimit,
-      template: ( item: EventItem ) => {
-        const eventContainer = document.createElement( this.settings.notePreviewOnHover ? 'a' : 'div' )
-        if ( 'href' in eventContainer ) {
-          eventContainer.addClass( 'internal-link' )
-          eventContainer.href = item.path
-        }
-
-        eventContainer.setText( item.content )
-
-        return eventContainer
-      }
-    }
-
-    timelineDiv.setAttribute( 'class', 'timeline-vis' )
-    const timeline = new Timeline( timelineDiv, items, options )
-
-    // these are probably non-performant but it works so ¯\_(ツ)_/¯
-    // dynamically add and remove a "special" class on hover
-    // cannot use standard :hover styling due to the structure
-    // of the timeline being so broken up across elements. This
-    // ensures that all elements related to an event are highlighted.
-    timeline.on( 'itemover', ( props ) => {
-      const event = items.get( props.item ) as unknown as EventItem
-      const newClass = event.className + ' runtime-hover'
-      document.documentElement.style.setProperty( '--hoverHighlightColor', event._event?.color ?? 'white' )
-      items.updateOnly( [{ ...event, className: newClass }] )
-
-      return () => {
-        timeline.off( 'itemover' )
-      }
-    })
-
-    timeline.on( 'itemout', ( props ) => {
-      const event = items.get( props.item ) as unknown as EventItem
-      const newClass = event.className.split( ' runtime-hover' )[0]
-      items.updateOnly( [{ ...event, className: newClass }] )
-
-      return () => {
-        timeline.off( 'itemout' )
-      }
-    })
-
-    // Replace the selected tags with the timeline html
-    el.appendChild( timelineDiv )
-  }
-
-  async showEmptyTimelineMessage( el: HTMLElement, tagList: string[] ) {
-    const timelineDiv = document.createElement( 'div' )
-    timelineDiv.setAttribute( 'class', 'empty-timeline' )
-    const message = `No events found for tags: [ '${tagList.join( "', '" )}' ]`
-
-    timelineDiv.createEl( 'p', { text: message })
-    el.appendChild( timelineDiv )
-  }
+  
 
   async run(
     source: string,
     el: HTMLElement,
-  ) {
+  ): Promise<void> {
     this.setup()
 
     // read arguments
@@ -570,7 +207,7 @@ export class TimelineBlockProcessor {
 
     if ( !this.currentFileList || this.currentFileList.length === 0 ) {
       logger( 'No files found for the timeline' )
-      await this.showEmptyTimelineMessage( el, Array.from( this.args.tags ))
+      await showEmptyTimelineMessage( el, Array.from( this.args.tags ))
       return
     }
 
@@ -588,10 +225,10 @@ export class TimelineBlockProcessor {
 
     switch ( this.args.type ) {
     case 'flat':
-      await this.buildHorizontalTimeline( timelineDiv, timelineNotes, sortedTimelineDates, el )
+      await buildHorizontalTimeline( timelineDiv, timelineNotes, sortedTimelineDates, el )
       return
     default:
-      await this.buildVerticalTimeline( timelineDiv, timelineNotes, sortedTimelineDates, el )
+      await buildVerticalTimeline( timelineDiv, timelineNotes, sortedTimelineDates, el )
       return
     }
   }
